+++ apps/dash-food-consumption/data/age.csv
@@ -1,0 +1,17 @@
+ages,percent
+5->9,7.9
+10->14,5.5
+15->19,6.3
+20->24,12.9
+25->29,18.3
+30->34,15
+35->39,11
+40->44,6.8
+45->49,5
+50->54,3.1
+55->59,2.9
+60->64,2.4
+65->69,1.3
+70->74,0.8
+75->79,0.3
+80+,0.5
