+++ apps/dash-food-consumption/data/gender.csv
@@ -1,0 +1,3 @@
+group,male,female
+migrate,48.2,51.8
+stay,49.9,50.1
