+++ apps/dash-food-consumption/data/income.csv
@@ -1,0 +1,64 @@
+Dia phuong,Thu_nhap_2021
+An Giang,3406.1
+B?c Giang,3965.7
+B?c K?n,2125.2
+B?c Liêu,3642.1
+B?c Ninh,4916.7
+B?n Tre,3367.3
+Bà R?a - Vtng Tàu,4419.5
+Bình  Duong,7123
+Bình Ð?nh,3468.5
+Bình Phu?c,4002.3
+Bình Thu?n,4076.9
+C?n Tho,4794.2
+Cà Mau,3239.3
+Cao B?ng,2273.3
+Ð?k L?k,2810.6
+Ð?k Nông,2745
+Ð?ng Nai,5751.5
+Ð?ng Tháp,4198.9
+Ðà N?ng,5229.9
+Ði?n Biên,1820.7
+Gia Lai,2329.4
+H?i Duong,4303.6
+H?i Phòng,5093.4
+H?u Giang,3761
+Hà Giang,1933.3
+Hà N?i,6001.6
+Hà Nam,4372.1
+Hà Tinh,2980.5
+Hòa Bình,2644.3
+Hung Yên,4192.3
+Khánh  Hòa,3235.5
+Kiên  Giang,3986.4
+Kon Tum,2517.4
+L?ng Son,2471.2
+Lai Châu,2049.6
+Lâm Ð?ng,3718.2
+Lào Cai,2515.5
+Long An,3724.9
+Nam Ð?nh,4412.5
+Ngh? An,3094.8
+Ninh  Thu?n,2855
+Ninh Bình,4281.7
+Phú Th?,3588.6
+Phú Yên,3296.4
+Qu?ng  Nam,3653.2
+Qu?ng  Ngãi,3219.5
+Qu?ng Bình,3338.2
+Qu?ng Ninh,3991.6
+Qu?ng Tr?,3025.7
+Sóc Trang,3246.2
+Son La,1834.3
+Tây Ninh,4036.4
+Th?a Thiên-Hu?,3525.4
+Thái  Nguyên,3813.7
+Thái Bình,4333.9
+Thanh Hóa,3651.6
+Ti?n Giang,3932.1
+TP. H? Chí Minh,6007.6
+Trà Vinh,3290.1
+Tuyên Quang,2872.4
+Vinh Long,3172.8
+Vinh Phúc,4511.2
+Yên Bái,2539.9
