+++ apps/dash-food-consumption/data/marriage.csv
@@ -1,0 +1,6 @@
+marriage_status,percent
+not married,22.9
+married,68.6
+widow/widower,6.3
+divorce,1.9
+separated,0.3
