+++ apps/dash-food-consumption/data/province_name.csv
@@ -1,0 +1,64 @@
+,Name
+0,An Giang
+1,Ba Ria - Vung Tau
+2,Bac Giang
+3,Bac Kan
+4,Bac Lieu
+5,Bac Ninh
+6,Ben Tre
+7,Binh Dinh
+8,Binh Duong
+9,Binh Phuoc
+10,Binh Thuan
+11,Ca Mau
+12,Can Tho
+13,Cao Bang
+14,Da Nang
+15,Dak Lak
+16,Dak Nong
+17,Dien Bien
+18,Dong Nai
+19,Dong Thap
+20,Gia Lai
+21,Ha Giang
+22,Ha Nam
+23,Ha Noi
+24,Ha Tinh
+25,Hai Duong
+26,Hai Phong
+27,Hau Giang
+28,Hoa Binh
+29,Hung Yen
+30,Khanh Hoa
+31,Kien Giang
+32,Kon Tum
+33,Lai Chau
+34,Lam Dong
+35,Lang Son
+36,Lao Cai
+37,Long An
+38,Nam Dinh
+39,Nghe An
+40,Ninh Binh
+41,Ninh Thuan
+42,Phu Tho
+43,Phu Yen
+44,Quang Binh
+45,Quang Nam
+46,Quang Ngai
+47,Quang Ninh
+48,Quang Tri
+49,Soc Trang
+50,Son La
+51,Tay Ninh
+52,Thai Binh
+53,Thai Nguyen
+54,Thanh Hoa
+55,Thua Thien - Hue
+56,Tien Giang
+57,TP. Ho Chi Minh
+58,Tra Vinh
+59,Tuyen Quang
+60,Vinh Long
+61,Vinh Phuc
+62,Yen Bai
