+++ apps/dash-food-consumption/data/reason.csv
@@ -1,0 +1,7 @@
+reason,percent
+new job,44.3
+losse job,0.5
+move new house with family,37.9
+marriage,7.2
+education,9
+others,1
