+++ apps/dash-food-footprint/app.py
@@ -1,717 +1,1000 @@
-import os
-
+import json
 import dash
 import dash_core_components as dcc
 import dash_html_components as html
-from dash.dependencies import Input, Output, State
-import dash_daq as daq
-import dash_bootstrap_components as dbc
-import numpy as np
+from dash.dependencies import Input, Output
+import plotly.express as px
+import urllib.request as urllib
+
+# from urllib.request import urlopen
 import pandas as pd
 import plotly.graph_objs as go
-import plotly.express as px
+import statsmodels.api as sm
+import plotly.figure_factory as ff
+# import geopandas as gpd
+import geojson
+import os
+# import numpy as np
+
+mapbox_access_token = "pk.eyJ1Ijoic3RlZmZlbmhpbGwiLCJhIjoiY2ttc3p6ODlrMG1ybzJwcG10d3hoaDZndCJ9.YE2gGNJiw6deBuFgHRHPjg"
+path = "https://raw.githubusercontent.com/FranzMichaelFrank/health_eu/main/"
+df = pd.read_csv(path + "food_supply.csv", dtype={"id": str})
+df_scatter = pd.read_csv(path + "scatter_data.csv", dtype={"id": str})
+clusters = pd.read_csv(path + "final_clusters.csv")
+box_cluster = pd.read_csv(path + "box_cluster.csv")
+
+url = path + "european-union-countries.geojson"
+response = urllib.urlopen(url)
+european_union = json.loads(response.read())
 
 dirname = os.path.dirname(__file__)
-path = os.path.join(dirname, "data/")
+path2 = os.path.join(dirname, "data/")
+# vn_map = gpd.read_file(path2 + "diaphantinhenglish.geojson", driver='GeoJSON')
+with open(path2 + "diaphantinhenglish.geojson") as f:
+    vn_map = geojson.load(f)
+vn_tctk = pd.read_csv(path2 + 'tctk.csv', sep=',')
 
-emissions = pd.read_csv(path + "emissions_with_origin.csv")
-productions = pd.read_csv(path + "productions.csv.gz")
-water = pd.read_csv(path + "water_use.csv")
-global_emissions = pd.read_csv(path + "Global_Emissions.csv")
-
-top10 = emissions.sort_values("Total_emissions")[-10:]
-top10_vegetal = emissions[emissions.Origin == "Vegetal"].sort_values("Total_emissions")[
-    -10:
-]
-top8_animal = emissions[emissions.Origin == "Animal"].sort_values("Total_emissions")
+def cz(s):
+    if s == "Czech Republic":
+        return "Czechia"
+    else:
+        return s
 
 
-radio_ani_veg = dbc.RadioItems(
-    id="ani_veg",
-    className="radio",
-    options=[
-        dict(label="Animal", value=0),
-        dict(label="Vegetal", value=1),
-        dict(label="Total", value=2),
-    ],
-    value=2,
-    inline=True,
-)
-
-dict_ = {
-    "Apples": "Apples",
-    "Bananas": "Bananas",
-    "Barley": "Barley",
-    "Beet Sugar": "Sugar beet",
-    "Berries & Grapes": "Berries & Grapes",
-    "Brassicas": "Brassicas",
-    "Cane Sugar": "Sugar cane",
-    "Cassava": "Cassava",
-    "Citrus Fruit": "Citrus",
-    "Coffee": "Coffee beans",
-    "Groundnuts": "Groundnuts",
-    "Maize": "Maize",
-    "Nuts": "Nuts",
-    "Oatmeal": "Oats",
-    "Olive Oil": "Olives",
-    "Onions & Leeks": "Onions & Leeks",
-    "Palm Oil": "Oil palm fruit",
-    "Peas": "Peas",
-    "Potatoes": "Potatoes",
-    "Rapeseed Oil": "Rapeseed",
-    "Rice": "Rice",
-    "Root Vegetables": "Roots and tubers",
-    "Soymilk": "Soybeans",
-    "Sunflower Oil": "Sunflower seed",
-    "Tofu": "Soybeans",
-    "Tomatoes": "Tomatoes",
-    "Wheat & Rye": "Wheat & Rye",
-    "Dark Chocolate": "Cocoa, beans",
-    "Milk": "Milk",
-    "Eggs": "Eggs",
-    "Poultry Meat": "Poultry Meat",
-    "Pig Meat": "Pig Meat",
-    "Seafood (farmed)": "Seafood (farmed)",
-    "Cheese": "Cheese",
-    "Lamb & Mutton": "Lamb & Mutton",
-    "Beef (beef herd)": "Beef (beef herd)",
-}
-
-options_veg = [
-    dict(label=key, value=dict_[key])
-    for key in top10_vegetal["Food product"].tolist()[::-1]
-    if key in dict_.keys()
-]
-options_an = [
-    dict(label=val, value=val) for val in top8_animal["Food product"].tolist()[::-1]
-]
-options_total = [
-    dict(label=key, value=dict_[key])
-    for key in top10["Food product"].tolist()[::-1]
-    if key in dict_.keys()
-]
-
-bar_colors = ["#ebb36a", "#6dbf9c"]
-bar_options = [top8_animal, top10_vegetal, top10]
-
-drop_map = dcc.Dropdown(
-    id="drop_map",
-    clearable=False,
-    searchable=False,
-    style={"margin": "4px", "box-shadow": "0px 0px #ebb36a", "border-color": "#ebb36a"},
-)
-
-drop_continent = dcc.Dropdown(
-    id="drop_continent",
-    clearable=False,
-    searchable=False,
-    options=[
-        {"label": "World", "value": "world"},
-        {"label": "Europe", "value": "europe"},
-        {"label": "Asia", "value": "asia"},
-        {"label": "Africa", "value": "africa"},
-        {"label": "North america", "value": "north america"},
-        {"label": "South america", "value": "south america"},
-    ],
-    value="world",
-    style={"margin": "4px", "box-shadow": "0px 0px #ebb36a", "border-color": "#ebb36a"},
-)
-
-slider_map = daq.Slider(
-    id="slider_map",
-    handleLabel={"showCurrentValue": True, "label": "Year"},
-    marks={str(i): str(i) for i in [1990, 1995, 2000, 2005, 2010, 2015]},
-    min=1990,
-    size=450,
-    color="#4B9072",
-)
-
-fig_water = px.sunburst(
-    water,
-    path=["Origin", "Category", "Product"],
-    values="Water Used",
-    color="Category",
-    color_discrete_sequence=px.colors.sequential.haline_r,
-).update_traces(hovertemplate="%{label}<br>" + "Water Used: %{value} L")
-
-fig_water = fig_water.update_layout(
-    {
-        "margin": dict(t=0, l=0, r=0, b=10),
-        "paper_bgcolor": "#F9F9F8",
-        "font_color": "#363535",
+clusters = clusters.rename(
+    columns={
+        "cluster_food": "Food clusters",
+        "cluster_health": "Health clusters",
+        "cluster_total": "Food & Health clusters",
     }
 )
+clusters["Food clusters"] = clusters["Food clusters"] + 1
+clusters["Food clusters"] = clusters["Food clusters"].apply(str)
+clusters["Health clusters"] = clusters["Health clusters"] + 1
+clusters["Health clusters"] = clusters["Health clusters"].apply(str)
+clusters["Food & Health clusters"] = clusters["Food & Health clusters"] + 1
+clusters["Food & Health clusters"] = clusters["Food & Health clusters"].apply(str)
 
-fig_gemissions = px.sunburst(
-    global_emissions,
-    path=["Emissions", "Group", "Subgroup"],
-    values="Percentage of food emissions",
-    color="Group",
-    color_discrete_sequence=px.colors.sequential.Peach_r,
-).update_traces(
-    hovertemplate="%{label}<br>" + "Global Emissions: %{value}%",
-    textinfo="label + percent entry",
+clusters["Country"] = clusters["Country"].apply(cz)
+df["Country"] = df["Country"].apply(cz)
+
+layout = dict(
+    autosize=True,
+    # automargin=True,
+    margin=dict(l=30, r=30, b=20, t=40),
+    hovermode="closest",
+    plot_bgcolor="#F9F9F9",
+    paper_bgcolor="#F9F9F9",
+    legend=dict(font=dict(size=10), orientation="h"),
+    title="Satellite Overview",
+    mapbox=dict(
+        accesstoken=mapbox_access_token,
+        style="light",
+        center=dict(lon=-78.05, lat=42.54),
+        zoom=7,
+    ),
 )
 
-fig_gemissions = fig_gemissions.update_layout(
-    {
-        "margin": dict(t=0, l=0, r=0, b=10),
-        "paper_bgcolor": "#F9F9F8",
-        "font_color": "#363535",
+# design for mapbox
+bgcolor = "#f3f3f1"  # mapbox light map land color
+row_heights = [150, 500, 300]
+template = {"layout": {"paper_bgcolor": bgcolor, "plot_bgcolor": bgcolor}}
+
+
+def blank_fig(height):
+    """
+    Build blank figure with the requested height
+    """
+    return {
+        "data": [],
+        "layout": {
+            "height": height,
+            "template": template,
+            "xaxis": {"visible": False},
+            "yaxis": {"visible": False},
+        },
     }
+
+
+columns = df.iloc[:, 2:22].columns
+health_cols = [
+    "Obesity",
+    "Diabetes Prevalence",
+    "Cardiovascular Death Rate",
+    "Life Expectancy",
+    "Health Expenditure",
+]
+
+# Create controls
+behaviour_options = [dict(label=country, value=country) for country in columns]
+
+
+food_options_ = [
+    "nhap_cu_2021",
+    "xuat_cu_2021",
+    "di_cu_2021",
+    "thu_nhap_2021",
+    "du_an_DTNN_2021",
+    "von_DTNN_2021",
+    "SV_DH_2020",
+    "SV_nghe_2020",
+    # "Offals",
+    # "Oilcrops",
+    # "Pulses",
+    # "Spices",
+    # "Starchy Roots",
+    # "Stimulants",
+    # "Sugar & Sweeteners",
+    # "Treenuts",
+    # "Vegetable Oils",
+    # "Vegetables",
+]
+
+food_options = [dict(label=country, value=country) for country in food_options_]
+
+
+dropdown_behaviour = dcc.Dropdown(
+    id="candidate_radio", options=behaviour_options, value=columns[0]
+)
+
+ots = [
+    "Alcoholic Beverages",
+    "Animal fats",
+    "Cereals - Excluding Beer",
+    "Eggs",
+    "Fish, Seafood",
+    "Fruits - Excluding Wine",
+    "Meat",
+    "Milk - Excluding Butter",
+    "Offals",
+    "Oilcrops",
+    "Pulses",
+    "Spices",
+    "Starchy Roots",
+    "Stimulants",
+    "Sugar & Sweeteners",
+    "Treenuts",
+    "Vegetable Oils",
+    "Vegetables",
+    "Obesity",
+    "Diabetes Prevalence",
+    "Cardiovascular Death Rate",
+    "Life Expectancy",
+    "Health Expenditure",
+]
+ots_ = [dict(label=country, value=country) for country in ots]
+
+ots_behaviour = dcc.Dropdown(id="box_dd", options=ots_, value="Alcoholic Beverages")
+
+radio_food_behaviour = dcc.RadioItems(
+    id="nutrition_types",
+    options=food_options,
+    value="nhap_cu_2021",
+    labelStyle={"display": "block", "text-align": "justify"},
 )
 
 
-# ------------------------------------------------------ APP ------------------------------------------------------
+corr_options = [
+    dict(label=country, value=country)
+    for country in [
+        "Obesity",
+        "Diabetes Prevalence",
+        "Cardiovascular Death Rate",
+        "Life Expectancy",
+        "Health Expenditure",
+    ]
+]
+cor_behav = dcc.Dropdown(
+    id="cor_behave",
+    options=corr_options,
+    value="Obesity"  # ,
+    # labelStyle={'display': 'block', "text-align": "justify"}
+)
+
+clust_opt = ["Food clusters", "Health clusters", "Food & Health clusters"]
+clust_options = [dict(label=country, value=country) for country in clust_opt]
+radio_clust_behaviour = dcc.RadioItems(
+    id="clust_types",
+    options=clust_options,
+    value="Food clusters",
+    labelStyle={"display": "inline-flex", "text-align": "center"},
+    style={"padding-left": "10%"},
+)
 
 app = dash.Dash(__name__)
 
-server = app.server
+###
+df_new = pd.read_csv("https://plotly.github.io/datasets/country_indicators.csv")
+available_indicators = df_new["Indicator Name"].unique()
+###
 
+
+## FF ##
+foods = [
+    "Alcoholic Beverages",
+    "Animal fats",
+    "Cereals - Excluding Beer",
+    "Eggs",
+    "Fish, Seafood",
+    "Fruits - Excluding Wine",
+    "Meat",
+    "Milk - Excluding Butter",
+    "Offals",
+    "Oilcrops",
+    "Pulses",
+    "Spices",
+    "Starchy Roots",
+    "Stimulants",
+    "Sugar & Sweeteners",
+    "Treenuts",
+    "Vegetable Oils",
+    "Vegetables",
+]
+
+foods_r = [
+    "Vegetables",
+    "Vegetable Oils",
+    "Treenuts",
+    "Sugar & Sweeteners",
+    "Stimulants",
+    "Starchy Roots",
+    "Spices",
+    "Pulses",
+    "Oilcrops",
+    "Offals",
+    "Milk - Excluding Butter",
+    "Meat",
+    "Fruits - Excluding Wine",
+    "Fish, Seafood",
+    "Eggs",
+    "Cereals - Excluding Beer",
+    "Animal fats",
+    "Alcoholic Beverages",
+]
+foods_health = [
+    "Vegetables",
+    "Vegetable Oils",
+    "Treenuts",
+    "Sugar & Sweeteners",
+    "Stimulants",
+    "Starchy Roots",
+    "Spices",
+    "Pulses",
+    "Oilcrops",
+    "Offals",
+    "Milk - Excluding Butter",
+    "Meat",
+    "Fruits - Excluding Wine",
+    "Fish, Seafood",
+    "Eggs",
+    "Cereals - Excluding Beer",
+    "Animal fats",
+    "Alcoholic Beverages",
+    "Obesity",
+    "Diabetes Prevalence",
+    "Cardiovascular Death Rate",
+    "Life Expectancy",
+    "Health Expenditure",
+]
+
+df_corr_r = df_scatter[foods_health]
+df_corr_round = df_corr_r.corr()[["Obesity"]].T[foods_r].T.round(2)
+# , "Diabetes Prevalence", "Cardiovascular Death Rate", "Life Expectancy", "Health Expenditure"
+fig_cor = ff.create_annotated_heatmap(
+    z=df_corr_round.to_numpy(),
+    x=df_corr_round.columns.tolist(),
+    y=df_corr_round.index.tolist(),
+    zmax=1,
+    zmin=-1,
+    showscale=True,
+    hoverongaps=True,
+)
+fig_cor.update_layout(
+    legend=dict(orientation="h", yanchor="bottom", y=1.02, xanchor="right", x=1)
+)
+# fig_cor.update_layout(yaxis_tickangle=-45)
+fig_cor.update_layout(xaxis_tickangle=0)
+fig_cor.update_layout(title_text="", height=600)
+
+
+health = df_scatter[
+    [
+        "Country",
+        "Obesity",
+        "Diabetes Prevalence",
+        "Cardiovascular Death Rate",
+        "Life Expectancy",
+        "Health Expenditure",
+    ]
+]
+
+
+fig_bar = go.Figure()
+fig_bar.add_trace(
+    go.Bar(
+        x=health["Country"], y=health["Obesity"], name="Obesity", marker_color="#0d0887"
+    )
+)
+fig_bar.add_trace(
+    go.Bar(
+        x=health["Country"],
+        y=health["Diabetes Prevalence"],
+        name="Diabetes Prevalence",
+        marker_color="#7201a8",
+    )
+)
+
+fig_bar.add_trace(
+    go.Bar(
+        x=health["Country"],
+        y=health["Cardiovascular Death Rate"],
+        name="Cardiovascular Death Rate",
+        marker_color="#bd3786",
+    )
+)
+
+fig_bar.add_trace(
+    go.Bar(
+        x=health["Country"],
+        y=health["Life Expectancy"],
+        name="Life Expectancy",
+        marker_color="#ed7953",
+    )
+)
+
+fig_bar.add_trace(
+    go.Bar(
+        x=health["Country"],
+        y=health["Health Expenditure"],
+        name="Health Expenditure",
+        marker_color="#fdca26",
+    )
+)
+# Here we modify the tickangle of the xaxis, resulting in rotated labels.
+fig_bar.update_layout(barmode="group", xaxis_tickangle=-45)
+fig_bar.update_layout(plot_bgcolor="white")
+fig_bar.update_yaxes(showline=True, linewidth=2, linecolor="black", gridcolor="grey")
+fig_bar.update_xaxes(showline=True, linewidth=2, linecolor="black")
+
+
+## FF ##
+
+# Create app layout
 app.layout = html.Div(
     [
+        dcc.Store(id="aggregate_data"),
+        # empty Div to trigger javascript file for graph resizing
+        html.Div(id="output-clientside"),
         html.Div(
             [
-                html.H1(children="FOOD FOOTPRINT"),
-                html.Label(
-                    "We are interested in investigating the food products that have the biggest impact on environment. Here you can understand which are the products whose productions emit more greenhouse gases and associate this with each supply chain step, their worldwide productions, and the water use.",
-                    style={"color": "rgb(33 36 35)"},
+                html.Div(
+                    [
+                        html.Img(
+                            src=app.get_asset_url("Nova_IMS.png"),
+                            id="plotly-image",
+                            style={
+                                "height": "60px",
+                                "width": "auto",
+                                "margin-bottom": "25px",
+                            },
+                        )
+                    ],
+                    className="one-third column",
                 ),
-                html.Img(
-                    src=app.get_asset_url("supply_chain.png"),
-                    style={
-                        "position": "relative",
-                        "width": "180%",
-                        "left": "-83px",
-                        "top": "-20px",
-                    },
+                html.Div(
+                    [
+                        html.Div(
+                            [
+                                html.H4(
+                                    "Food consumption characteristics of the European Union",
+                                    style={"font-weight": "bold"},
+                                ),
+                                html.H5(
+                                    "Analysis of the relationship between nutritional patterns and \n the health status within the countries",
+                                    style={"margin-top": "0px"},
+                                ),
+                            ]
+                        )
+                    ],
+                    className="three column",
+                    id="title",
+                ),
+                html.Div(
+                    # create empty div for align cente
+                    className="one-third column",
                 ),
             ],
-            className="side_bar",
+            id="header",
+            className="row flex-display",
+            style={"margin-bottom": "25px"},
         ),
         html.Div(
             [
                 html.Div(
                     [
-                        html.Div(
-                            [
-                                html.Label("Choose the Product's Origin:"),
-                                html.Br(),
-                                html.Br(),
-                                radio_ani_veg,
-                            ],
-                            className="box",
+                        html.H6(
+                            "Consumption by food type",
                             style={
-                                "margin": "10px",
-                                "padding-top": "15px",
-                                "padding-bottom": "15px",
+                                "margin-top": "0",
+                                "font-weight": "bold",
+                                "text-align": "center",
                             },
                         ),
-                        html.Div(
-                            [
-                                html.Div(
-                                    [
-                                        html.Div(
-                                            [
-                                                html.Label(id="title_bar"),
-                                                dcc.Graph(id="bar_fig"),
-                                                html.Div(
-                                                    [html.P(id="comment")],
-                                                    className="box_comment",
-                                                ),
-                                            ],
-                                            className="box",
-                                            style={"padding-bottom": "15px"},
-                                        ),
-                                        html.Div(
-                                            [
-                                                html.Img(
-                                                    src=app.get_asset_url("Food.png"),
-                                                    style={
-                                                        "width": "100%",
-                                                        "position": "relative",
-                                                        "opacity": "80%",
-                                                    },
-                                                ),
-                                            ]
-                                        ),
-                                    ],
-                                    style={"width": "40%"},
-                                ),
-                                html.Div(
-                                    [
-                                        html.Div(
-                                            [
-                                                html.Label(
-                                                    id="choose_product",
-                                                    style={"margin": "10px"},
-                                                ),
-                                                drop_map,
-                                            ],
-                                            className="box",
-                                        ),
-                                        html.Div(
-                                            [
-                                                html.Div(
-                                                    [
-                                                        html.Label(
-                                                            "Emissions measured as kg of CO2 per kg of product",
-                                                            style={
-                                                                "font-size": "medium"
-                                                            },
-                                                        ),
-                                                        html.Br(),
-                                                        html.Br(),
-                                                        html.Div(
-                                                            [
-                                                                html.Div(
-                                                                    [
-                                                                        html.H4(
-                                                                            "Land use",
-                                                                            style={
-                                                                                "font-weight": "normal"
-                                                                            },
-                                                                        ),
-                                                                        html.H3(
-                                                                            id="land_use"
-                                                                        ),
-                                                                    ],
-                                                                    className="box_emissions",
-                                                                ),
-                                                                html.Div(
-                                                                    [
-                                                                        html.H4(
-                                                                            "Animal Feed",
-                                                                            style={
-                                                                                "font-weight": "normal"
-                                                                            },
-                                                                        ),
-                                                                        html.H3(
-                                                                            id="animal_feed"
-                                                                        ),
-                                                                    ],
-                                                                    className="box_emissions",
-                                                                ),
-                                                                html.Div(
-                                                                    [
-                                                                        html.H4(
-                                                                            "Farm",
-                                                                            style={
-                                                                                "font-weight": "normal"
-                                                                            },
-                                                                        ),
-                                                                        html.H3(
-                                                                            id="farm"
-                                                                        ),
-                                                                    ],
-                                                                    className="box_emissions",
-                                                                ),
-                                                                html.Div(
-                                                                    [
-                                                                        html.H4(
-                                                                            "Processing",
-                                                                            style={
-                                                                                "font-weight": "normal"
-                                                                            },
-                                                                        ),
-                                                                        html.H3(
-                                                                            id="processing"
-                                                                        ),
-                                                                    ],
-                                                                    className="box_emissions",
-                                                                ),
-                                                                html.Div(
-                                                                    [
-                                                                        html.H4(
-                                                                            "Transport",
-                                                                            style={
-                                                                                "font-weight": "normal"
-                                                                            },
-                                                                        ),
-                                                                        html.H3(
-                                                                            id="transport"
-                                                                        ),
-                                                                    ],
-                                                                    className="box_emissions",
-                                                                ),
-                                                                html.Div(
-                                                                    [
-                                                                        html.H4(
-                                                                            "Packaging",
-                                                                            style={
-                                                                                "font-weight": "normal"
-                                                                            },
-                                                                        ),
-                                                                        html.H3(
-                                                                            id="packging"
-                                                                        ),
-                                                                    ],
-                                                                    className="box_emissions",
-                                                                ),
-                                                                html.Div(
-                                                                    [
-                                                                        html.H4(
-                                                                            "Retail",
-                                                                            style={
-                                                                                "font-weight": "normal"
-                                                                            },
-                                                                        ),
-                                                                        html.H3(
-                                                                            id="retail"
-                                                                        ),
-                                                                    ],
-                                                                    className="box_emissions",
-                                                                ),
-                                                            ],
-                                                            style={"display": "flex"},
-                                                        ),
-                                                    ],
-                                                    className="box",
-                                                    style={"heigth": "10%"},
-                                                ),
-                                                html.Div(
-                                                    [
-                                                        html.Div(
-                                                            [
-                                                                html.Div(
-                                                                    [
-                                                                        html.Br(),
-                                                                        html.Label(
-                                                                            id="title_map",
-                                                                            style={
-                                                                                "font-size": "medium"
-                                                                            },
-                                                                        ),
-                                                                        html.Br(),
-                                                                        html.Label(
-                                                                            "These quantities refer to the raw material used to produce the product selected above",
-                                                                            style={
-                                                                                "font-size": "9px"
-                                                                            },
-                                                                        ),
-                                                                    ],
-                                                                    style={
-                                                                        "width": "70%"
-                                                                    },
-                                                                ),
-                                                                html.Div(
-                                                                    [],
-                                                                    style={
-                                                                        "width": "5%"
-                                                                    },
-                                                                ),
-                                                                html.Div(
-                                                                    [
-                                                                        drop_continent,
-                                                                        html.Br(),
-                                                                        html.Br(),
-                                                                    ],
-                                                                    style={
-                                                                        "width": "25%"
-                                                                    },
-                                                                ),
-                                                            ],
-                                                            className="row",
-                                                        ),
-                                                        dcc.Graph(
-                                                            id="map",
-                                                            style={
-                                                                "position": "relative",
-                                                                "top": "-50px",
-                                                            },
-                                                        ),
-                                                        html.Div(
-                                                            [slider_map],
-                                                            style={
-                                                                "margin-left": "15%",
-                                                                "position": "relative",
-                                                                "top": "-38px",
-                                                            },
-                                                        ),
-                                                    ],
-                                                    className="box",
-                                                    style={"padding-bottom": "0px"},
-                                                ),
-                                            ]
-                                        ),
-                                    ],
-                                    style={"width": "60%"},
-                                ),
-                            ],
-                            className="row",
+                        html.P(
+                            "The cultures and customs of the 27 EU countries differ widely. The same applies to their eating and drinking habits."
+                            " The map on the right explores the food supply in kilograms per capita per year.",
+                            className="control_label",
+                            style={"text-align": "justify"},
                         ),
-                        html.Div(
-                            [
-                                html.Div(
-                                    [
-                                        html.Label(
-                                            "3. Global greenhouse gas emissions from food production, in percentage",
-                                            style={"font-size": "medium"},
-                                        ),
-                                        html.Br(),
-                                        html.Label(
-                                            "Click on it to know more!",
-                                            style={"font-size": "9px"},
-                                        ),
-                                        html.Br(),
-                                        html.Br(),
-                                        dcc.Graph(figure=fig_gemissions),
-                                    ],
-                                    className="box",
-                                    style={"width": "40%"},
-                                ),
-                                html.Div(
-                                    [
-                                        html.Label(
-                                            "4. Freshwater withdrawals per kg of product, in Liters",
-                                            style={"font-size": "medium"},
-                                        ),
-                                        html.Br(),
-                                        html.Label(
-                                            "Click on it to know more!",
-                                            style={"font-size": "9px"},
-                                        ),
-                                        html.Br(),
-                                        html.Br(),
-                                        dcc.Graph(figure=fig_water),
-                                    ],
-                                    className="box",
-                                    style={"width": "63%"},
-                                ),
-                            ],
-                            className="row",
+                        html.P(),
+                        html.P(
+                            "Select a food category",
+                            className="control_label",
+                            style={"text-align": "center", "font-weight": "bold"},
                         ),
+                        radio_food_behaviour,
+                    ],
+                    className="pretty_container four columns",
+                    id="cross-filter-options",
+                    style={"text-align": "justify"},
+                ),
+                html.Div(
+                    [
                         html.Div(
-                            [
-                                html.Div(
-                                    [
-                                        html.P(
-                                            [
-                                                "GroupV",
-                                                html.Br(),
-                                                "Ana Carrelha (20200631), Inês Melo (20200624), Inês Roque (20200644), Ricardo Nunes(20200611)",
-                                            ],
-                                            style={"font-size": "12px"},
-                                        ),
-                                    ],
-                                    style={"width": "60%"},
-                                ),
-                                html.Div(
-                                    [
-                                        html.P(
-                                            [
-                                                "Sources ",
-                                                html.Br(),
-                                                html.A(
-                                                    "Our World in Data",
-                                                    href="https://ourworldindata.org/",
-                                                    target="_blank",
-                                                ),
-                                                ", ",
-                                                html.A(
-                                                    "Food and Agriculture Organization of the United Nations",
-                                                    href="http://www.fao.org/faostat/en/#data",
-                                                    target="_blank",
-                                                ),
-                                            ],
-                                            style={"font-size": "12px"},
-                                        )
-                                    ],
-                                    style={"width": "37%"},
-                                ),
-                            ],
-                            className="footer",
-                            style={"display": "flex"},
+                            [dcc.Graph(id="choropleth")],
+                            # id="countGraphContainer",
+                            className="pretty_container",
                         ),
                     ],
-                    className="main",
+                    id="right-column",
+                    className="eight columns",
                 ),
-            ]
+            ],
+            className="row flex-display",
         ),
-    ]
+        html.Div(
+            [
+                html.H6(
+                    "General health information about the countries",
+                    style={
+                        "margin-top": "0",
+                        "font-weight": "bold",
+                        "text-align": "center",
+                    },
+                ),
+                html.P(
+                    "Similarly to nutrition, the health status also varies from country to country. The bar chart below shows the differences between the countries in terms of the following variables: prevalence of obesity in the adult population in % (Obesity), prevalence of diabetes in the adult population in % (Diabetes Prevalence), cardiovascular death rate per 100,000 population (Cardiovascular Death Rate), average life expectancy in years (Life Expectancy) and the expenditure of the government on the country's health system in % of the respective GDP (Health Expenditure).",
+                    className="control_label",
+                    style={"text-align": "justify"},
+                ),
+                html.Div(
+                    [dcc.Graph(id="bar_chart", figure=fig_bar)],
+                    className="pretty_container twelve columns",
+                ),
+            ],
+            className="row pretty_container",
+        ),
+        html.Div(
+            [
+                html.Div(
+                    # cor_behav,
+                    [
+                        html.H6(
+                            "Exploring correlations",
+                            style={
+                                "margin-top": "0",
+                                "font-weight": "bold",
+                                "text-align": "center",
+                            },
+                        ),
+                        html.P(
+                            "In the heatmap below, the correlations between the 5 health variables and the 18 food variables can be explored.",
+                            className="control_label",
+                            style={"text-align": "justify"},
+                        ),
+                        # html.P("""<br>"""),
+                        html.P(
+                            "Select a health variable",
+                            style={"font-weight": "bold", "text-align": "center"},
+                        ),
+                        cor_behav,
+                        html.Div(
+                            [dcc.Graph(id="cor_ma")],
+                            className="pretty_container twelve columns",
+                        ),
+                    ],  # ,cor_behav,
+                    className="pretty_container four columns",
+                ),
+                html.Div(
+                    [
+                        html.H6(
+                            "Analysing the correlations between food consumption and health",
+                            style={
+                                "margin-top": "0",
+                                "font-weight": "bold",
+                                "text-align": "center",
+                            },
+                        ),
+                        html.P(
+                            "Below, the correlations can be analysed in more detail. It is important to note that correlation in this case does not necessarily mean causation. For example, the wealth level of a country might often influence the variables, which is why it is represented through the size of the dots as GDP per capita. Furthermore, outliers should also be watched out for.",
+                            className="control_label",
+                            style={"text-align": "justify"},
+                        ),
+                        html.Div(
+                            [
+                                html.P(
+                                    "Select a food category",
+                                    className="control_label",
+                                    style={
+                                        "font-weight": "bold",
+                                        "text-align": "center",
+                                    },
+                                ),
+                                dcc.Dropdown(
+                                    id="xaxis-column",
+                                    options=[
+                                        {"label": i, "value": i} for i in food_options_
+                                    ],
+                                    value="Alcoholic Beverages",  # ,className="pretty_container four columns",
+                                ),
+                                dcc.RadioItems(
+                                    id="xaxis-type",
+                                    options=[
+                                        {"label": i, "value": i}
+                                        for i in ["Box", "Violin"]
+                                    ],
+                                    value="Box",
+                                    labelStyle={
+                                        "display": "inline-block"
+                                    },  # ,className="pretty_container four columns",
+                                    style={"padding-left": "34%"},
+                                ),
+                            ],
+                            className="pretty_container sixish columns",
+                        ),
+                        html.Div(
+                            [
+                                html.P(
+                                    "Select a health variable",
+                                    className="control_label",
+                                    style={
+                                        "font-weight": "bold",
+                                        "text-align": "center",
+                                    },
+                                ),
+                                dcc.Dropdown(
+                                    id="yaxis-column",
+                                    options=[
+                                        {"label": i, "value": i} for i in health_cols
+                                    ],
+                                    value=health_cols[0],
+                                ),
+                                dcc.RadioItems(
+                                    id="yaxis-type",
+                                    options=[
+                                        {"label": i, "value": i}
+                                        for i in ["Box", "Violin"]
+                                    ],
+                                    value="Box",
+                                    labelStyle={"display": "inline-block"},
+                                    style={"padding-left": "34%"},
+                                ),
+                            ],
+                            className="pretty_container sixish columns",
+                        ),
+                        html.Div(
+                            [dcc.Graph(id="indicator-graphic"),],
+                            className="pretty_container almost columns",
+                        ),
+                    ],
+                    className="pretty_container eight columns",
+                ),
+            ],
+            className="row flex-display",
+        ),
+        html.Div(
+            [
+                html.H6(
+                    "K-means clustering",
+                    style={
+                        "margin-top": "0",
+                        "font-weight": "bold",
+                        "text-align": "center",
+                    },
+                ),
+                html.P(
+                    "Finally, k-means clustering is carried out. The criterion can be selected on the left side, whereby either the 18 food variables, the 5 health variables or all of them in combination may be chosen for the clustering. On the right side, the resulting clusters can then be compared with respect to a chosen variable.",
+                    className="control_label",
+                    style={"text-align": "justify"},
+                ),
+                html.Div(
+                    [
+                        html.P(
+                            "Select a clustering criterion",
+                            className="control_label",
+                            style={"text-align": "center", "font-weight": "bold"},
+                        ),
+                        radio_clust_behaviour,
+                        dcc.Graph(id="cluster_map"),
+                    ],
+                    className="pretty_container sixish columns",
+                ),
+                html.Div(
+                    [
+                        html.P(
+                            "Select a variable for cluster comparison",
+                            className="control_label",
+                            style={"text-align": "center", "font-weight": "bold"},
+                        ),
+                        ots_behaviour,
+                        dcc.Graph(id="boxes"),
+                    ],
+                    className="pretty_container sixish columns",
+                ),
+            ],
+            className="row pretty_container",
+        ),
+        html.Div(
+            [
+                html.H6(
+                    "Authors",
+                    style={
+                        "margin-top": "0",
+                        "font-weight": "bold",
+                        "text-align": "center",
+                    },
+                ),
+                html.P(
+                    "Maximilian Maukner (m20200645@novaims.unl.pt)  -  Ehsan Meisami Fard (m20201050@novaims.unl.pt)  -  Franz Michael Frank (m20200618@novaims.unl.pt)  -  Steffen Hillmann (m20200589@novaims.unl.pt)",
+                    style={"text-align": "center", "font-size": "10pt"},
+                ),
+            ],
+            className="row pretty_container",
+        ),
+        html.Div(
+            [
+                html.H6(
+                    "Sources",
+                    style={
+                        "margin-top": "0",
+                        "font-weight": "bold",
+                        "text-align": "center",
+                    },
+                ),
+                dcc.Markdown(
+                    """\
+                         - Eurostat: https://ec.europa.eu/eurostat/databrowser/view/HLTH_SHA11_HF__custom_227597/bookmark/table?lang=en&bookmarkId=1530a1e6-767e-4661-9e15-0ed2f7fae0d5
+                         - Food and Agriculture Organization of the United Nations: http://www.fao.org/faostat/en/#data/FBS
+                         - Opendatasoft: https://data.opendatasoft.com/explore/dataset/european-union-countries@public/export/
+                         - Our World in Data: https://covid.ourworldindata.org/data/owid-covid-data.csv?v=2021-03-11
+                        """,
+                    style={"font-size": "10pt"},
+                ),
+            ],
+            className="row pretty_container",
+        ),
+    ],
+    id="mainContainer",
+    style={"display": "flex", "flex-direction": "column"},
 )
 
+colors = [
+    "#0d0887",
+    "#46039f",
+    "#7201a8",
+    "#9c179e",
+    "#bd3786",
+    "#d8576b",
+    "#ed7953",
+    "#fb9f3a",
+    "#fdca26",
+    "#f0f921",
+]
+colors2 = ["#fdca26", "#ed7953", "#bd3786", "#7201a8", "#0d0887"]
 
-# ------------------------------------------------------ Callbacks ------------------------------------------------------
+####
+@app.callback(Output("choropleth", "figure"), [Input("nutrition_types", "value")])
+def display_choropleth(candi):
+    fig = px.choropleth_mapbox(
+        vn_tctk,
+        geojson=vn_map,
+        color=candi,
+        locations="Name",
+        featureidkey="properties.Name",
+        hover_name="Name",
+        opacity=0.7,  # hover_data = [],
+        center={"lat": 15, "lon": 106},
+        zoom=4,
+    )
+    fig.update_layout(
+        margin={"r": 0, "t": 0, "l": 0, "b": 0}, mapbox_accesstoken=mapbox_access_token
+    )
+
+    return fig
+
+
+@app.callback(
+    Output("boxes", "figure"), [Input("clust_types", "value"), Input("box_dd", "value")]
+)
+def display_boxes(clust, vari):
+    clusts = ["Food", "Health", "Food & Health"]
+
+    box_cluster = pd.read_csv(path + "box_cluster.csv").rename(
+        columns={
+            "Food": "Food clusters",
+            "Health": "Health clusters",
+            "Food & Health": "Food & Health clusters",
+        }
+    )
+    box_cluster["Food clusters"] = box_cluster["Food clusters"] + 1
+    box_cluster["Health clusters"] = box_cluster["Health clusters"] + 1
+    box_cluster["Food & Health clusters"] = box_cluster["Food & Health clusters"] + 1
+    box_cluster = box_cluster.sort_values(clust)
+    fig = px.box(
+        box_cluster,
+        x=clust,
+        y=vari,
+        color=clust,
+        color_discrete_sequence=["#fdca26", "#ed7953", "#bd3786", "#7201a8", "#0d0887"],
+    )
+    fig.update_layout(plot_bgcolor="white")
+    fig.update_yaxes(showline=True, linewidth=2, linecolor="black")
+    fig.update_xaxes(showline=True, linewidth=2, linecolor="black")
+    fig.update_layout(legend=dict(yanchor="bottom", y=0))
+
+    return fig
+
+
+@app.callback(Output("cluster_map", "figure"), [Input("clust_types", "value")])
+def display_cluster_map(type_clust):
+    fig = px.choropleth_mapbox(
+        clusters.sort_values(type_clust),
+        locations="id",
+        geojson=european_union,
+        featureidkey="properties.gu_a3",
+        color=type_clust,
+        hover_name="Country",
+        hover_data=[type_clust],
+        # hover_data=["Life Expectancy"],
+        # title="clusters by food",
+        mapbox_style="carto-positron",
+        center={"lat": 56.5, "lon": 11},
+        zoom=2.5,
+        opacity=0.7,
+        color_discrete_sequence=colors2,
+    )
+
+    fig.update_layout(
+        legend=dict(yanchor="bottom", y=0),
+        margin={"r": 0, "t": 0, "l": 0, "b": 0},
+        mapbox_accesstoken=mapbox_access_token,
+    )
+
+    return fig
+
+
+## FF ##
+@app.callback(Output("cor_ma", "figure"), [Input("cor_behave", "value")])
+def display_cor_ma(var):
+    foods = [
+        "Alcoholic Beverages",
+        "Animal fats",
+        "Cereals - Excluding Beer",
+        "Eggs",
+        "Fish, Seafood",
+        "Fruits - Excluding Wine",
+        "Meat",
+        "Milk - Excluding Butter",
+        "Offals",
+        "Oilcrops",
+        "Pulses",
+        "Spices",
+        "Starchy Roots",
+        "Stimulants",
+        "Sugar & Sweeteners",
+        "Treenuts",
+        "Vegetable Oils",
+        "Vegetables",
+    ]
+
+    foods_r = [
+        "Vegetables",
+        "Vegetable Oils",
+        "Treenuts",
+        "Sugar & Sweeteners",
+        "Stimulants",
+        "Starchy Roots",
+        "Spices",
+        "Pulses",
+        "Oilcrops",
+        "Offals",
+        "Milk - Excluding Butter",
+        "Meat",
+        "Fruits - Excluding Wine",
+        "Fish, Seafood",
+        "Eggs",
+        "Cereals - Excluding Beer",
+        "Animal fats",
+        "Alcoholic Beverages",
+    ]
+    foods_health = [
+        "Vegetables",
+        "Vegetable Oils",
+        "Treenuts",
+        "Sugar & Sweeteners",
+        "Stimulants",
+        "Starchy Roots",
+        "Spices",
+        "Pulses",
+        "Oilcrops",
+        "Offals",
+        "Milk - Excluding Butter",
+        "Meat",
+        "Fruits - Excluding Wine",
+        "Fish, Seafood",
+        "Eggs",
+        "Cereals - Excluding Beer",
+        "Animal fats",
+        "Alcoholic Beverages",
+        "Obesity",
+        "Diabetes Prevalence",
+        "Cardiovascular Death Rate",
+        "Life Expectancy",
+        "Health Expenditure",
+    ]
+
+    df_corr_r = df_scatter[foods_health]
+    df_corr_round = df_corr_r.corr()[[var]].T[foods_r].T.round(2)
+    # , "Diabetes Prevalence", "Cardiovascular Death Rate", "Life Expectancy", "Health Expenditure"
+    fig_cor = ff.create_annotated_heatmap(
+        z=df_corr_round.to_numpy(),
+        x=df_corr_round.columns.tolist(),
+        y=df_corr_round.index.tolist(),
+        zmax=1,
+        zmin=-1,
+        showscale=True,
+        hoverongaps=True,
+        ygap=3,
+    )
+    fig_cor.update_layout(
+        yaxis=dict(showgrid=False),
+        xaxis=dict(showgrid=False),
+        legend=dict(orientation="h", yanchor="bottom", y=1.02, xanchor="right", x=1),
+    )
+    # fig_cor.update_layout(yaxis_tickangle=-45)
+    fig_cor.update_layout(xaxis_tickangle=0)
+    fig_cor.update_layout(title_text="", height=550)  #
+
+    return fig_cor
+
+
+## FF ##
 
 
 @app.callback(
     [
-        Output("title_bar", "children"),
-        Output("bar_fig", "figure"),
-        Output("comment", "children"),
-        Output("drop_map", "options"),
-        Output("drop_map", "value"),
-        Output("choose_product", "children"),
+        Output("max_name", "children"),
+        Output("max_value", "children"),
+        Output("min_name", "children"),
+        Output("min_value", "children"),
+        Output("mean", "children"),
+        Output("st_dev", "children"),
     ],
-    [Input("ani_veg", "value")],
+    [Input("nutrition_types", "value"),],
 )
-def bar_chart(top10_select):
+def indicator(auswahl):
+    max_id = df[auswahl].idxmax()
+    min_id = df[auswahl].idxmin()
 
-    ################## Top10 Plot ##################
-    title = "1. Greenhouse emissions (kg CO2 per kg of product)"
-    df = bar_options[top10_select]
+    max_value = df[auswahl].max()
+    max_value = str(max_value)
 
-    if top10_select == 2:
-        bar_fig = dict(
-            type="bar",
-            x=df.Total_emissions,
-            y=df["Food product"],
-            orientation="h",
-            marker_color=["#ebb36a" if x == "Animal" else "#6dbf9c" for x in df.Origin],
-        )
-    else:
-        bar_fig = dict(
-            type="bar",
-            x=df.Total_emissions,
-            y=df["Food product"],
-            orientation="h",
-            marker_color=bar_colors[top10_select],
-        )
+    max_name = df.loc[max_id, "Country"]
+    min_value = df[auswahl].min()
+    min_value = str(min_value)
 
-    ################## Dropdown Bar ##################
-    if top10_select == 0:
-        options_return = options_an
-        product_chosen = "2. Choose an animal product:"
-        comment = [
-            "Look at the beef production emissions! Each kilogram of beef produces almost 60 kg of CO2.",
-            html.Br(),
-            html.Br(),
-        ]
-    elif top10_select == 1:
-        options_return = options_veg
-        product_chosen = "2. Choose a vegetal product:"
-        comment = [
-            "Did you know that dark chocolate and coffee are the vegetal-based products that emit more gases?",
-            html.Br(),
-            html.Br(),
-        ]
-    else:
-        options_return = options_total
-        product_chosen = "2. Choose an animal or vegetal product:"
-        comment = "Check the difference between animal and vegetal-based products! Beef (top1 animal-based emitter) produces around 3 times more emissions than dark chocolate (top1 plant-based emitter)."
+    min_name = df.loc[min_id, "Country"]
+    mean = df[auswahl].mean()
+    st_dev = df[auswahl].std()
+    st_dev = round(st_dev, 2)
+    st_dev = str(st_dev)
+    mean = round(mean, 2)
+    mean = str(mean)
 
     return (
-        title,
-        go.Figure(
-            data=bar_fig,
-            layout=dict(
-                height=300,
-                font_color="#363535",
-                paper_bgcolor="rgba(0,0,0,0)",
-                plot_bgcolor="rgba(0,0,0,0)",
-                margin=dict(l=20, r=20, t=30, b=20),
-                margin_pad=10,
-            ),
-        ),
-        comment,
-        options_return,
-        options_return[0]["value"],
-        product_chosen,
+        "Country: " + max_name,
+        max_value + " kg per capita per year",
+        "Country: " + min_name,
+        min_value + " kg per capita per year",
+        mean + " kg per capita per year",
+        st_dev + " kg per capita per year",
     )
 
 
 @app.callback(
-    [Output("slider_map", "max"), Output("slider_map", "value"),],
-    [Input("drop_map", "value")],
+    Output("indicator-graphic", "figure"),
+    Input("xaxis-column", "value"),
+    Input("yaxis-column", "value"),
+    Input("xaxis-type", "value"),
+    Input("yaxis-type", "value"),
 )
-def update_slider(product):
-    year = productions[productions["Item"] == product]["Year"].max()
-    return year, year
+def update_graph(xaxis_column_name, yaxis_column_name, xaxis_type, yaxis_type):
 
-
-@app.callback(
-    [
-        Output("land_use", "children"),
-        Output("animal_feed", "children"),
-        Output("farm", "children"),
-        Output("processing", "children"),
-        Output("transport", "children"),
-        Output("packging", "children"),
-        Output("retail", "children"),
-        Output("title_map", "children"),
-        Output("map", "figure"),
-    ],
-    [
-        Input("drop_map", "value"),
-        Input("slider_map", "value"),
-        Input("drop_continent", "value"),
-    ],
-    [State("drop_map", "options")],
-)
-def update_map(drop_map_value, year, continent, opt):
-
-    ################## Emissions datset ##################
-
-    the_label = [x["label"] for x in opt if x["value"] == drop_map_value]
-
-    data_emissions = emissions[emissions["Food product"] == the_label[0]]
-    land_use_str = str(np.round(data_emissions["Land use change"].values[0], 2))
-    animal_feed_str = str(np.round(data_emissions["Animal Feed"].values[0], 2))
-    farm_str = str(np.round(data_emissions["Farm"].values[0], 2))
-    processing_str = str(np.round(data_emissions["Processing"].values[0], 2))
-    transport_str = str(np.round(data_emissions["Transport"].values[0], 2))
-    packging_str = str(np.round(data_emissions["Packging"].values[0], 2))
-    retail_str = str(np.round(data_emissions["Retail"].values[0], 2))
-
-    ################## Choroplet Plot ##################
-
-    prod1 = productions[
-        (productions["Item"] == drop_map_value) & (productions["Year"] == year)
-    ]
-
-    title = "Production quantities of {}, by country".format(
-        prod1["Item"].unique()[0]
-    )  # font_color = '#363535',
-    data_slider = []
-    data_each_yr = dict(
-        type="choropleth",
-        locations=prod1["Area"],
-        locationmode="country names",
-        autocolorscale=False,
-        z=np.log(prod1["Value"].astype(float)),
-        zmin=0,
-        zmax=np.log(productions[productions["Item"] == drop_map_value]["Value"].max()),
-        colorscale=["#ffe2bd", "#006837"],
-        marker_line_color="rgba(0,0,0,0)",
-        colorbar={"title": "Tonnes (log)"},  # Tonnes in logscale
-        colorbar_lenmode="fraction",
-        colorbar_len=0.8,
-        colorbar_x=1,
-        colorbar_xanchor="left",
-        colorbar_y=0.5,
-        name="",
-    )
-    data_slider.append(data_each_yr)
-
-    layout = dict(
-        geo=dict(
-            scope=continent,
-            projection={"type": "natural earth"},
-            bgcolor="rgba(0,0,0,0)",
-        ),
-        margin=dict(l=0, r=0, b=0, t=30, pad=0),
-        paper_bgcolor="rgba(0,0,0,0)",
-        plot_bgcolor="rgba(0,0,0,0)",
+    # col_name = str(yaxis_column_name) + " (above Average)"
+    col_name = " "
+    df_scatter[col_name] = (
+        df_scatter[yaxis_column_name] > df_scatter[yaxis_column_name].mean()
     )
 
-    fig_choropleth = go.Figure(data=data_slider, layout=layout)
-    fig_choropleth.update_geos(
-        showcoastlines=False, showsubunits=False, showframe=False
+    def aa(inp):
+        if inp == True:
+            return yaxis_column_name + " above average"
+        else:
+            return yaxis_column_name + " below average"
+
+    df_scatter[col_name] = df_scatter[col_name].apply(aa)
+
+    if yaxis_type == "Box":
+        type_y = "box"
+    else:
+        type_y = "violin"
+
+    if xaxis_type == "Box":
+        type_x = "box"
+    else:
+        type_x = "violin"
+
+    fig = px.scatter(
+        df_scatter,
+        x=xaxis_column_name,
+        y=yaxis_column_name,
+        size="GDP per Capita",
+        color=col_name,
+        hover_name="Country",
+        log_x=False,
+        marginal_x=type_x,
+        marginal_y=type_y,
+        template="simple_white",
+        color_discrete_sequence=["#0d0887", "#9c179e"],
     )
 
-    return (
-        land_use_str,
-        animal_feed_str,
-        farm_str,
-        processing_str,
-        transport_str,
-        packging_str,
-        retail_str,
-        title,
-        fig_choropleth,
+    # linear regression
+    regline = (
+        sm.OLS(
+            df_scatter[yaxis_column_name],
+            sm.add_constant(df_scatter[xaxis_column_name]),
+        )
+        .fit()
+        .fittedvalues
     )
 
+    # add linear regression line for whole sample
+    fig.add_traces(
+        go.Scatter(
+            x=df_scatter[xaxis_column_name],
+            y=regline,
+            mode="lines",
+            marker_color="#fb9f3a",
+            name="OLS Trendline",
+        )
+    )
+
+    fig.update_layout(
+        legend=dict(orientation="h", xanchor="center", x=0.5, yanchor="top", y=-0.2)
+    )
+
+    fig.update_layout(margin={"l": 40, "b": 40, "t": 10, "r": 0}, hovermode="closest")
+
+    # fig.update_xaxes(title=xaxis_column_name,
+    #                type='linear' if xaxis_type == 'Linear' else 'log')
+
+    # fig.update_yaxes(title=yaxis_column_name,
+    #                type='linear' if yaxis_type == 'Linear' else 'log')
+
+    return fig
+
+
+server = app.server
 
 if __name__ == "__main__":
     app.run_server(debug=True)
