+++ apps/dash-food-footprint/data/school.csv
@@ -1,0 +1,3 @@
+Ages,at shool,stop learning,never went to school
+6->10,97.6,1.4,1
+11->18,55.7,43.8,0.5
