+++ apps/dash-food-footprint/data/tctk.csv
@@ -1,0 +1,64 @@
+Name,nhap_cu_2021,xuat_cu_2021,di_cu_2021,thu_nhap_2021,du_an_DTNN_2021,von_DTNN_2021,SV_DH_2020,SV_nghe_2020
+An Giang,2.93,25.79,-22.85,3406.1,29,292.81,10203,24436
+Bac Giang,4.16,3.53,0.63,3965.7,568,8618.03,1192,31567
+Bac Kan,1.42,13.43,-12.01,2125.2,4,7.9,,6274
+Bac Lieu,0.31,17.91,-17.61,3642.1,14,4551.01,1654,9273
+Bac Ninh,45.66,3.93,41.73,4916.7,1717,22485.18,7727,67116
+Ben Tre,3.04,8.39,-5.35,3367.3,64,1583.94,,9415
+Ba Ria - Vung Tau,4.85,3.93,0.92,4419.5,517,33014.45,4841,34862
+Binh Dinh,2.33,7.04,-4.7,3468.5,97,1188.29,20294,29941
+Binh Duong,35.63,4.87,30.75,7123,4022,37791.63,24652,47894
+Binh Phuoc,1.95,4.17,-2.22,4002.3,379,3732.44,,37396
+Binh Thuan,0.56,5.91,-5.35,4076.9,155,3822.27,2267,16516
+Can Tho,12.46,13.36,-0.91,4794.2,84,2052.91,84717,61560
+Ca Mau,0.82,20.22,-19.39,3239.3,11,157.83,,10849
+Cao Bang,2.5,7.29,-4.79,2273.3,17,36.42,,5550
+Dak Lak,1.18,6.89,-5.7,2810.6,25,635.89,11766,32889
+Dak Nong,4.61,8.12,-3.51,2745,20,311.38,,5829
+Dong Nai,9.5,8.8,0.6,5751.5,1796,33999.29,22452,84105
+Dong Thap,1.79,25.48,-23.69,4198.9,22,232.81,11804,16618
+Da Nang,18.93,5.41,13.51,5229.9,889,5932.41,99573,40570
+Dien Bien,0.69,7.36,-6.66,1820.7,1,3,,8921
+Gia Lai,0.49,5.89,-5.4,2329.4,7,21.09,,16331
+Hai Duong,1.01,3.03,-2.02,4303.6,493,8886.09,11449,38116
+Hai Phong,5.93,2.03,3.9,5093.4,897,23609.13,40989,68115
+Hau Giang,1.08,18.81,-17.73,3761,29,686.08,4253,9729
+Ha Giang,0.32,7.89,-7.57,1933.3,6,4.15,,10130
+Ha Noi,7.68,1.75,5.93,6001.6,6700,37582.11,628981,265723
+Ha Nam,3.01,4.52,-1.51,4372.1,358,4894.45,,18939
+Ha Tinh,2.74,7.93,-5.19,2980.5,79,11739.24,1399,31071
+Hoa Binh,2,12.35,-10.4,2644.3,51,720.14,,15305
+Hung Yen,3.64,3.7,-0.07,4192.3,515,6099.47,13507,30524
+Khanh Hoa,0.67,2.69,-2.02,3235.5,118,4394.33,19688,29815
+Kien Giang,1.46,16.86,-15.4,3986.4,62,4808.15,4486,25970
+Kon Tum,1.4,5.84,-4.43,2517.4,10,246.72,1386,14815
+Lang Son,0.86,13.77,-12.91,2471.2,42,240.36,,19195
+Lai Chau,0.81,3.87,-3.06,2049.6,1,1.5,,9546
+Lam Dong,2.07,1.99,0.09,3718.2,102,530.24,17069,28802
+Lao Cai,0.69,6.75,-6.06,2515.5,32,582.63,,10593
+Long An,8.67,6.54,2.13,3724.9,1253,12270.71,2406,28200
+Nam Dinh,1.04,9.36,-8.33,4412.5,123,3659.89,18112,37920
+Nghe An,1.38,7.57,-6.19,3094.8,107,1977.36,33364,76804
+Ninh Binh,2.39,4.69,-2.3,4281.7,89,1586.33,1136,25049
+Ninh Thuan,2.04,3.77,-1.73,2855,55,1728.3,,11586
+Phu Tho,1.68,2.41,-0.73,3588.6,211,2613.81,7534,32354
+Phu Yen,2.34,5.24,-2.91,3296.4,53,2035.96,4275,10651
+Quang Binh,2.66,5.4,-2.74,3338.2,24,1116.28,1888,19626
+Quang Nam,3.46,8.14,-4.67,3653.2,221,6058.36,3164,35220
+Quang Ngai,2.49,5.47,-2.99,3219.5,59,2029.42,3064,25319
+Quang Ninh,0.37,3.44,-3.07,3991.6,152,7949.98,4038,32625
+Quang Tri,3.48,6.65,-3.17,3025.7,26,2526.81,135,16045
+Soc Trang,0.93,21.8,-20.9,3246.2,14,228.78,,12830
+Son La,0.27,5.08,-4.81,1834.3,10,135.73,4488,26172
+Tay Ninh,7.35,3.63,3.73,4036.4,350,8503.61,,15808
+Thua Thien - Hue,6.02,3.88,2.14,3525.4,124,4041.7,55382,30831
+Thai Binh,0.71,6.22,-5.51,4333.9,104,1399.06,8168,22728
+Thai Nguyen,1.74,4.47,-2.73,3813.7,193,9840.97,47232,45067
+Thanh Hoa,0.69,7.85,-7.16,3651.6,168,14715.32,10472,50152
+Tien Giang,3.71,8.73,-5.02,3932.1,128,2885.04,4782,13631
+TP. Ho Chi Minh,25.4,3.59,21.8,6007.6,10394,52921.58,599782,337637
+Tra Vinh,1.64,12.85,-11.21,3290.1,43,3344.31,20818,21516
+Tuyen Quang,3.29,12.2,-8.92,2872.4,18,208.82,2846,9167
+Vinh Long,3.1,10.28,-7.18,3172.8,67,874.73,15530,33400
+Vinh Phuc,2.31,2.14,0.17,4511.2,477,6518.24,10650,43818
+Yen Bai,1.54,8.08,-6.54,2539.9,33,448.56,,22015
