+++ apps/dash-food-footprint/data/training.csv
@@ -1,0 +1,6 @@
+group,percent
+no professional knowledge,62.8
+elementary level,7.6
+intermediate level,4.6
+college degree,5.9
+university degree and above,19.1
